requirements.txt
aiohttp==3.7.4.post0
anyascii==0.2.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
bitcoin==1.1.42
bleach==3.3.0
blockchain==1.4.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
commerce==0.1.dev0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.5
django-503==0.1
django-appconf==1.0.4
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-commerce==0.1.0
django-common-helpers==0.9.2
django-contact==0.1.1
django-contact-us==0.4.4
django-countries==7.2.1
django-crispy-forms==1.12.0
django-cron==0.5.1
django-crontab==0.7.1
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-debug-toolbar-django13==0.8.4
django-django-csv-exports==1.0.0
django-email==0.1.10
django-environ==0.4.4
django-etc==1.3.0
django-filebrowser-django13==3.0
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-iban==0.3.1
django-import-export==2.5.0
django-invoicing==3.1.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modelcluster==5.1
django-money==2.0.1
django-notifications-hq==1.6.0
django-notifier==0.7
django-phonenumber-field
django-recaptcha==2.0.6
django-resized==0.3.11
django-simple-contact==1.0.0
django-simple-referrals==0.2
django-storages==1.11.1
django-taggit==1.5.1
django-tinymce==3.3.0
django-tools-contact==1.3.2
django-tracking-analyzer==1.1.1
django-tracking-analyzer-django2==0.3
django-treebeard==4.5.1
django-user-agents==0.4.0
django-utils-six==2.0
django-websites==0.0.1
django-zotero==0.4
djangorestframework==3.12.4
draftjs-exporter==2.1.7
dropbox==11.13.2
enum-compat==0.0.3
et-xmlfile==1.1.0
Flask==2.0.1
future==0.18.2
geoip2==4.2.0
giant-contact==0.6
giant-mixins==0.3
google-api-core==1.31.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonfield==3.1.0
l18n==2020.6.1
MarkupPy==1.14
MarkupSafe==2.0.1
maxminddb==2.0.3
multidict==5.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
paypal==1.2.5
phonenumbers==8.12.28
Pillow==8.3.1
pinax-referrals==4.0.1
ply==3.11
protobuf==3.17.3
psycopg2==2.9.1
py-moneyed==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-magic==0.4.24
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rest-framework-pagination==0.0.1
rsa==4.7.2
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stone==3.2.1
stripe==2.59.0
swapper==1.1.2.post1
sweetify==2.0.0
tablib==3.0.0
telepath==0.2
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
vatnumber==1.2
wagtail==2.13.3
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.4
xlwt==1.3.0
yarl==1.6.3
